requirement.txt
a2a-sdk==0.3.7
absolufy-imports==0.3.1
ag-ui-adk==0.3.1
ag-ui-protocol==0.1.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
asyncio==4.0.0
attrs==25.3.0
authlib==1.6.4
backoff==2.2.1
bcrypt==5.0.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
choreographer==1.1.1
chromadb==1.1.0
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
cryptography==46.0.1
db-dtypes==1.4.3
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
docstring-parser==0.17.0
durationpy==0.10
executing==2.2.1
fastapi==0.117.1
filelock==3.19.1
flasgger==0.9.7.1
flask==3.1.2
flask-sock==0.7.0
flatbuffers==25.9.23
frozenlist==1.7.0
fsspec==2025.9.0
google-adk==1.14.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.115.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.38.0
google-cloud-bigtable==2.32.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-spanner==3.58.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.38.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpc-interceptor==0.15.4
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython-pygments-lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.6.3
jupyter-core==5.8.1
kaleido==1.1.0
kubernetes==34.1.0
logistro==1.1.0
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mcp==1.14.1
mdurl==0.1.2
mistune==3.1.4
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
narwhals==2.6.0
nest-asyncio==1.6.0
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.23.0
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
posthog==5.4.0
prompt-toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
psycopg2-binary==2.9.10
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyodbc==5.2.0
pyparsing==3.2.5
pypika==0.48.9
pyproject-hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pyyaml==6.0.2
pyzmq==27.1.0
quickchart-io==2.0.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sqlalchemy-spanner==1.16.0
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tenacity==8.5.0
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.37.0
vanna==0.7.9
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
